# frontend/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Login API",
    "version": "1.0.0",
    "description": "Login system APIs"
  },
  "servers": [
    {
      "url": "http://localhost:3001/api"
    }
  ],
  "paths": {
    "/login": {
      "post": {
        "summary": "Login to your account",
        "tags": [
          "auth-controller"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "You are logged",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "406": {
            "description": "Wrong email or password"
          }
        }
      }
    },
    "/register": {
      "post": {
        "summary": "Create a new account",
        "tags": [
          "auth-controller"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The user was successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterRequest"
                }
              }
            }
          },
          "500": {
            "description": "Some server error"
          }
        }
      }
    },
    "/role/update": {
      "put": {
        "summary": "Update role user",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "id",
            "schema": {
              "type": "integer",
              "description": "Id role"
            }
          },
          {
            "in": "query",
            "name": "email",
            "schema": {
              "type": "string",
              "description": "The email of the user to whom we are changing the role"
            }
          }
        ],
        "tags": [
          "user-controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "aplication/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        }
      }
    },
    "/user": {
      "get": {
        "summary": "Return the data from the pointed email",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "email",
            "schema": {
              "type": "string",
              "description": "The current page number"
            }
          }
        ],
        "tags": [
          "user-controller"
        ],
        "responses": {
          "200": {
            "description": "The list of the users",
            "content": {
              "aplication/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        }
      }
    },
    "/add-friend": {
      "post": {
        "summary": "Add friend route",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "friendId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "tags": [
          "friend-controller"
        ],
        "responses": {
          "200": {
            "description": "Friend request has been sent successfully"
          }
        }
      }
    },
    "/accept-friend": {
      "post": {
        "summary": "Accept friend route",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "friendId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "tags": [
          "friend-controller"
        ],
        "responses": {
          "200": {
            "description": "Friend request has been accepted successfully"
          },
          "409": {
            "description": "Friend request already accepted"
          }
        }
      }
    },
    "/reject-friend": {
      "post": {
        "summary": "Reject friend request",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "friendId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "tags": [
          "friend-controller"
        ],
        "responses": {
          "200": {
            "description": "Friend request rejected successfully"
          }
        }
      }
    },
    "/friends-list": {
      "get": {
        "summary": "Lists of all friends",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "friend-controller"
        ],
        "responses": {
          "200": {
            "description": "List of all friends",
            "content": {
              "aplication/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArrayOfFriends"
                }
              }
            }
          }
        }
      }
    },
    "/add-friend-list": {
      "get": {
        "summary": "Lists of all unfriends",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "friend-controller"
        ],
        "responses": {
          "200": {
            "description": "List of all unfriends",
            "content": {
              "aplication/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArrayOfFriends"
                }
              }
            }
          }
        }
      }
    },
    "/get-all-conversations": {
      "get": {
        "summary": "Get all users conversations",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "friend-controller"
        ],
        "responses": {
          "200": {
            "description": "List of all conversations",
            "content": {
              "aplication/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArrayOfConversations"
                }
              }
            }
          }
        }
      }
    },
    "/get-messages": {
      "get": {
        "summary": "Get all messages from an conversations",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "chatId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "chat-controller"
        ],
        "responses": {
          "200": {
            "description": "List of all messages",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArrayOfMessages"
                }
              }
            }
          }
        }
      }
    },
    "/users": {
      "get": {
        "summary": "Return the list of all users",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "description": "The current page number"
            }
          }
        ],
        "tags": [
          "user-controller"
        ],
        "responses": {
          "200": {
            "description": "The list of the users",
            "content": {
              "aplication/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableDataDto"
                }
              }
            }
          }
        }
      }
    },
    "/user/current": {
      "get": {
        "summary": "Return the curent user data.",
        "parameters": [
          {
            "in": "header",
            "name": "token",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "tags": [
          "user-controller"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "aplication/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update user data",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDto"
              }
            }
          }
        },
        "tags": [
          "user-controller"
        ],
        "responses": {
          "200": {
            "description": "The user data has been updated successfully"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "UserDto": {
        "type": "object",
        "required": [
          "lastName",
          "firstName",
          "email",
          "city",
          "state",
          "country",
          "profilePictureUrl"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "minLength": 1,
            "description": "The unique id of each user",
            "readOnly": true
          },
          "firstName": {
            "type": "string",
            "minLength": 1,
            "description": "The first name of each user"
          },
          "lastName": {
            "type": "string",
            "minLength": 1,
            "description": "The last name of each user"
          },
          "email": {
            "type": "string",
            "minLength": 1,
            "description": "The email of each user"
          },
          "city": {
            "type": "string",
            "minLength": 1,
            "description": "The city of each user"
          },
          "state": {
            "type": "string",
            "minLength": 1,
            "description": "The gender for each user"
          },
          "country": {
            "type": "string",
            "minLength": 1,
            "description": "The country for each user"
          },
          "profilePictureUrl": {
            "type": "string",
            "minLength": 1,
            "description": "The profile picture URL"
          },
          "userRole": {
            "$ref": "#/components/schemas/RoleDto"
          }
        },
        "example": {
          "id": 0,
          "firstName": "string",
          "lastName": "string",
          "email": "string",
          "city": "string",
          "state": "string",
          "country": "string",
          "profilePictureUrl": "string",
          "userRole": {
            "id": 0,
            "name": "ADMIN"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "minLength": 1
          }
        },
        "example": {
          "email": "string",
          "password": "string"
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": [
          "firstName",
          "lastName",
          "email",
          "password"
        ],
        "properties": {
          "firstName": {
            "type": "string",
            "minLength": 1
          },
          "lastName": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "minLength": 1
          }
        },
        "example": {
          "firstName": "string",
          "lastName": "string",
          "email": "string",
          "password": "string"
        }
      },
      "RoleDto": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "enum": [
              "ADMIN",
              "SUPERADMIN",
              "USER"
            ]
          }
        }
      },
      "TableDataDto": {
        "required": [
          "page"
        ],
        "properties": {
          "page": {
            "type": "integer",
            "minLength": 1
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataDto"
            }
          }
        },
        "example": {
          "page": 1,
          "data": {
            "id": 0,
            "firstName": "string",
            "lastName": "string",
            "email": "string",
            "created_at": "string",
            "profilePictureUrl": "string",
            "userRole": {
              "id": 0,
              "role": "SUPERADMIN"
            }
          }
        }
      },
      "DataDto": {
        "required": [
          "id",
          "firstName",
          "lastName",
          "email",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "minLength": 1
          },
          "firstName": {
            "type": "string",
            "minLength": 1
          },
          "lastName": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string",
            "minLength": 1
          },
          "created_at": {
            "type": "string",
            "minLength": 1
          },
          "userRole": {
            "$ref": "#/components/schemas/RoleDto"
          }
        },
        "example": {
          "id": 0,
          "firstName": "string",
          "lastName": "string",
          "email": "string",
          "created_at": "string",
          "userRole": {
            "id": 0,
            "role": "SUPERADMIN"
          }
        }
      },
      "FriendDto": {
        "required": [
          "email",
          "firstName",
          "id",
          "lastName",
          "status"
        ],
        "properties": {
          "email": {
            "type": "string",
            "minLength": 1
          },
          "firstName": {
            "type": "string",
            "minLength": 1
          },
          "id": {
            "type": "integer",
            "minLength": 1
          },
          "lastName": {
            "type": "string"
          },
          "profilePictureUrl": {
            "type": "string",
            "minLength": 1
          },
          "status": {
            "type": "string",
            "minLength": 1
          },
          "chatId": {
            "type": "integer",
            "minLength": 1
          }
        },
        "example": {
          "email": "string",
          "firstName": "string",
          "id": "integer",
          "lastName": "string",
          "profilePictureUrl": "string",
          "status": "string"
        }
      },
      "MessageDto": {
        "required": [
          "chatId",
          "createAt",
          "id",
          "lineText",
          "userId"
        ],
        "properties": {
          "chatId": {
            "type": "string",
            "minLength": 1
          },
          "createAt": {
            "type": "string",
            "minLength": 1
          },
          "id": {
            "type": "integer",
            "minLength": 1
          },
          "lineText": {
            "type": "string",
            "minLength": 1
          },
          "userId": {
            "type": "string",
            "minLength": 1
          }
        },
        "example": {
          "chatId": "string",
          "createAt": "string",
          "id": "string",
          "lineText": "string",
          "userId": "string"
        }
      },
      "ArrayOfMessages": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/MessageDto"
        }
      },
      "ArrayOfFriends": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/FriendDto"
        }
      },
      "ConversationDto": {
        "required": [
          "chatId",
          "lastLineText",
          "lastSentMessage"
        ],
        "properties": {
          "chatId": {
            "type": "integer",
            "minLength": 1
          },
          "firstName": {
            "type": "string",
            "minLength": 1
          },
          "lastLineText": {
            "type": "string",
            "minLength": 1
          },
          "lastSentMessage": {
            "type": "string",
            "minLength": 1
          },
          "lastName": {
            "type": "string",
            "minLength": 1
          },
          "groupName": {
            "type": "string",
            "minLength": 1
          },
          "groupPicture": {
            "type": "string",
            "minLength": 1
          },
          "profilePictureUrl": {
            "type": "string",
            "minLength": 1
          }
        },
        "example": {
          "chatId": "integer",
          "firstName": "string",
          "lastLineText": "string",
          "lastSentMessage": "string",
          "lastName": "string"
        }
      },
      "ArrayOfConversations": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/ConversationDto"
        }
      }
    }
  },
  "tags": [
    {
      "name": "auth-controller"
    },
    {
      "name": "user-controller"
    },
    {
      "name": "chat-controller"
    },
    {
      "name": "friend-controller"
    }
  ]
}